pyproject.toml
[project]
name = "support"
version = "0.1.0"
description = "A FastAPI-based chatbot using Google Gemini for conversational AI and SQLAlchemy (Core) for chat persistence "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.15.0",
    "sqlalchemy[asyncio]>=2.0.41",
]
